% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSOdata.R
\name{readSOdata}
\alias{readSOdata}
\title{Read data from Stack Overflow easily}
\usage{
readSOdata(text, sep = "|", strip.white = TRUE, header = TRUE, ...)
}
\arguments{
\item{text}{string to be read}

\item{sep}{separator}

\item{strip.white}{Do you want to strip the white space?}

\item{header}{Does the text contain headers?}

\item{...}{Additional arguments to read.table}
}
\value{
A data.frame
}
\description{
Read data from Stack Overflow easily
}
\examples{
readSOdata('Participate | B1  | B2  | B3  | B4  | B5  | Query1  | Query2
3           | -1  | -1  | -1  | -1  | -1  | Noissue | Noissue
1           | -1  | 1   | -1  | -1  | 1   | Noissue | Noissue
1           | -1  | -1  | -1  | -1  | -1  | Issue   | Noissue
2           | -1  | 1   | 1   | -1  | 1   | Noissue | Noissue
2           | 1   | 1   | 1   | 1   | -1  | Noissue | Noissue
1           | -99 | -99 | -99 | -99 | -99 | Noissue | Noissue')
}
